PDJB-794: propagate tag and flag changes
