+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/summaryModels/propertyComplianceViewModels/PropertyComplianceViewModelFactoryTests.kt
@@ -35,7 +35,7 @@ class PropertyComplianceViewModelFactoryTests {
         PropertyComplianceViewModelFactory(
             gasSafetyViewModelFactory,
             electricalSafetyViewModelFactory,
-            EpcViewModelServiceRedesign(mock()),
+            EpcViewModelFactory(mock()),
             epcCertificateUrlProvider,
             NotificationBannerViewModelServiceRedesign(),
         )
